# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface Appointment {
  id: number;
  clientName: string;
  phone: string;
  service: string;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  /** Duration in minutes */
  duration: number;
  price: number;
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
  updatedAt: string;
}